Eusuite&DeploymentNando/apps/eusuite-company-backend/requirements.txt
# Company Admin Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1
pillow==10.2.0
boto3==1.34.0
kubernetes==28.1.0
pyyaml==6.0.1
jinja2==3.1.3
redis==5.0.1
celery==5.3.4
python-slugify==8.0.1
